fix: make mocks in separate index-file so they are not mixed with implementation
