+++ packages/suite/src/actions/wallet/__tests__/signVerifyActions.test.ts
@@ -1,4 +1,4 @@
-import { mockSuiteDevice } from '@suite-common/suite-types';
+import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { testMocks } from '@suite-common/test-utils';
 
 import { showAddress, sign, verify } from 'src/actions/wallet/signVerifyActions';
