+++ suite-common/suite-types/mocks/index.ts
@@ -1,0 +1,5 @@
+export {
+    mockConnectDevice,
+    mockSuiteDevice,
+    mockGetFirmwareReleaseConfigInfo,
+} from './mockSuiteDevice';
