+++ suite-common/suite-types/mocks/mockSuiteDevice.ts
