+++ suite-common/suite-types/src/index.ts
@@ -16,9 +16,3 @@ export type Selector<TReturnValue> = (state: any) => TReturnValue;
 export type SuiteCompatibleAction<TPayload> = (
     payload: TPayload,
 ) => AnyAction | ActionCreatorWithPayload<TPayload> | ActionCreatorWithoutPayload;
-
-export {
-    mockConnectDevice,
-    mockSuiteDevice,
-    mockGetFirmwareReleaseConfigInfo,
-} from '../tests/mockSuiteDevice';
