+++ suite-common/suite-utils/src/__fixtures__/device.ts
@@ -1,4 +1,5 @@
-import { TrezorDevice, mockSuiteDevice } from '@suite-common/suite-types';
+import { TrezorDevice } from '@suite-common/suite-types';
+import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { DeviceModelInternal } from '@trezor/device-utils';
 import * as URLS from '@trezor/urls';
