+++ suite-common/test-utils/src/extraDependenciesCommonMock.ts
@@ -14,7 +14,8 @@ import {
     notImplementedThunk,
 } from '@suite-common/redux-utils';
 import type { SuiteSync } from '@suite-common/suite-sync-types';
-import { ReportSecurityCheckParams, Route, mockSuiteDevice } from '@suite-common/suite-types';
+import { ReportSecurityCheckParams, Route } from '@suite-common/suite-types';
+import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { AddressDisplayOptions, SelectedAccountLoaded } from '@suite-common/wallet-types';
 import { Analytics } from '@trezor/analytics-uploader';
 import { err, ok } from '@trezor/type-utils';
