+++ suite-common/thp/tests/autoInitThpAfterDeviceConnectionThunk.test.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
 import { FirmwareUpdateState, prepareFirmwareReducer } from '@suite-common/firmware';
-import { mockConnectDevice } from '@suite-common/suite-types';
+import { mockConnectDevice } from '@suite-common/suite-types/mocks';
 import { configureMockStore, extraDependenciesCommonMock } from '@suite-common/test-utils';
 import { acquireDevice, prepareDeviceReducer } from '@suite-common/wallet-core';
 import { Device } from '@trezor/connect';
