+++ suite-common/wallet-types/src/index.ts
@@ -17,4 +17,3 @@ export * from './globalSendReceive';
 export * from './device';
 export * from './baseCurrency';
 export * from './transactionSimulation';
-
