+++ suite-common/wallet-utils/src/__tests__/accountUtils.test.ts
@@ -1,4 +1,4 @@
-import { mockSuiteDevice } from '@suite-common/suite-types';
+import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { testMocks } from '@suite-common/test-utils';
 import { NetworkFeature } from '@suite-common/wallet-config';
 import { Account, asAccountDescriptor } from '@suite-common/wallet-types';
