+++ suite-native/device-authorization/src/__tests__/deviceAuthorizationSlice.test.ts
@@ -1,4 +1,4 @@
-import { mockSuiteDevice } from '@suite-common/suite-types';
+import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { UI } from '@trezor/connect';
 
 import {
